move two more rules to the top level
